backend/requirements.txt
celery[redis]==5.2.1
cryptography==36.0.0
django==4.0.0
django-ipware==4.0.2
djangorestframework==3.12.4
drf_yasg==1.21.4
gunicorn==20.1.0
nanoid==2.0.0
psycopg2-binary==2.9.2